requirements.txt
openai
langchain
langchain-openai
langchain_google_genai
langchain_community
anthropic
google-generativeai
perplexityai
groq
python-dotenv
playwright
selenium
webdriver-manager
notion-client
langchain_ollama
crewai
crewai[tools]
rich
gradio
